Threads can now be locked/stickied.
- Locking a thread prevents any updates, new posts, deleting posts, etc. from happening
- Making a thread sticky will make it appear at the top of a topic.

+++ k9r-backend/api/src/forum/delete.rs
@@ -1,17 +1,17 @@
-use actix_web::{web, HttpMessage, HttpRequest, HttpResponse, Responder};
+use actix_web::{web, HttpMessage, HttpRequest, HttpResponse};
 use k9r_db::{crud::{forum_posts::{delete_forum_post_from_id, get_forum_post_from_id, get_forum_posts_in_forum_thread}, forum_threads::{delete_forum_thread_from_id, get_forum_thread_from_id, update_forum_thread_from_id}}, models::User};
 
 use crate::models::Message;
 
 pub async fn delete_post(
     (request, path): (HttpRequest, web::Path<(i32, )>),
-) -> Result<impl Responder, Box<dyn std::error::Error>> {
+) -> HttpResponse {
     let user = match request.extensions().get::<User>().cloned() {
         Some(user) => user,
         None => {
-            return Ok(HttpResponse::Unauthorized().json(Message {
+            return HttpResponse::Unauthorized().json(Message {
                 message: "Failed to get user".to_string(),
-            }))
+            });
         }
     };
 
@@ -24,12 +24,18 @@ pub async fn delete_post(
                         true => {
                             let thread_option = get_forum_thread_from_id(post_id);
                             if thread_option.is_none() {
-                                return Ok(HttpResponse::NotFound().json(Message {
+                                return HttpResponse::NotFound().json(Message {
                                     message: "Failed to get thread".to_string()
-                                }));
+                                });
                             }
 
                             let mut thread = thread_option.unwrap();
+                            if thread.locked {
+                                return HttpResponse::BadRequest().json(Message { 
+                                    message: "Can't update a locked thread".to_string()
+                                });
+                            }
+
                             for (index, id) in thread.posts.iter().enumerate() {
                                 if id == &post.id {
                                     thread.posts.remove(index);
@@ -40,26 +46,26 @@ pub async fn delete_post(
                             let update = serde_json::from_str(serde_json::to_string(&thread).unwrap().as_str()).unwrap();
                             update_forum_thread_from_id(thread.id, update);
 
-                            Ok(HttpResponse::Ok().json(thread))
+                            HttpResponse::Ok().json(thread)
                         },
                         false => {
-                            Ok(HttpResponse::BadRequest().json(Message {
+                            HttpResponse::BadRequest().json(Message {
                                 message: "Failed to delete forum post".to_string()
-                            }))
+                            })
                         }
                     }
                 }
                 false => {
-                    Ok(HttpResponse::Unauthorized().json(Message {
+                    HttpResponse::Unauthorized().json(Message {
                         message: "Invalid user".to_string()
-                    }))
+                    })
                 }
             }
         }
         None => {
-            Ok(HttpResponse::NotFound().json(Message {
+            HttpResponse::NotFound().json(Message {
                 message: "Failed to get forum post".to_string()
-            }))
+            })
         }
     }
 }
@@ -79,6 +85,12 @@ pub async fn delete_thread(
     let thread_id = path.into_inner().0;
     match get_forum_thread_from_id(thread_id) {
         Some(thread) => {
+            if thread.locked {
+                return HttpResponse::BadRequest().json(Message { 
+                    message: "Can't update a locked thread".to_string()
+                });
+            }
+
             match thread.author == user.id {
                 true => {
                     match delete_forum_thread_from_id(thread.id) {
